connectors/GK Today Website Data Extraction as Connector/requirement.txt
requests==2.31.0
beautifulsoup4==4.12.2
mysql-connector-python==8.0.33
Pillow==10.0.1